Crate amethyst_gltf

Structs

GltfAnimation
A single animation

GltfMaterial
GLTF material, PBR based

GltfNode
A node in the scene hierarchy

GltfPrimitive
A single graphics primitive

GltfScene
A single scene is defined as a list of the root nodes in the node hierarchy for the full asset

GltfSceneAsset
Actual asset produced on finished loading of a GLTF scene file.

GltfSceneFormat
Gltf scene format, will cause the whole default scene to be loaded from the given file.

GltfSceneLoaderSystem
A GLTF scene loader, will transform Handle<GltfSceneAsset> into full entity hierarchies.

GltfSceneOptions
Options used when loading a GLTF file

GltfTexture
A GLTF defined texture, will be in TextureData format in the output from the loader.

Enums

AlphaMode
Alpha mode for material