Struct amethyst_gltf::GltfPrefab

pub struct GltfPrefab {
    pub transform: Option<Transform>,
    pub mesh: Option<MeshData>,
    pub mesh_handle: Option<Handle<Mesh>>,
    pub material: Option<MaterialPrefab<TextureFormat>>,
    pub animatable: Option<AnimatablePrefab<usize, Transform>>,
    pub skinnable: Option<SkinnablePrefab>,
    pub extent: Option<GltfNodeExtent>,
}

PrefabData for loading Gltf files.

Fields
transform: Option<Transform>

Transform will almost always be placed, the only exception is for the main Entity for certain scenarios (based on the data in the Gltf file)

mesh: Option<MeshData>

MeshData is placed on all Entitys with graphics primitives

mesh_handle: Option<Handle<Mesh>>

Mesh handle after sub asset loading is done

material: Option<MaterialPrefab<TextureFormat>>

MeshData is placed on all Entitys with graphics primitives with material

animatable: Option<AnimatablePrefab<usize, Transform>>

Loaded animations, if applicable, will always only be placed on the main Entity

skinnable: Option<SkinnablePrefab>

Skin data is placed on Entitys involved in the skin, skeleton or graphical primitives using the skin

extent: Option<GltfNodeExtent>

Node extent

Methods

impl GltfPrefab

pub fn move_to(&mut self, target: Point3<f32>)

Move the scene so the center of the bounding box is at the given target location.

pub fn scale_to(&mut self, max_distance: f32)

Scale the scene to a specific max size