Struct amethyst_gltf::GltfSceneFormat

pub struct GltfSceneFormat;

Gltf scene format, will cause the whole default scene to be loaded from the given file.

Using the GltfSceneLoaderSystem a Handle<GltfSceneAsset> from this format can be attached to an entity in ECS, and the system will then load the full scene using the given entity as the root node of the scene hierarchy.