Struct amethyst_gltf::GltfSceneFormat

pub struct GltfSceneFormat;

Gltf scene format, will load a single scene from a Gltf file.

Using the GltfSceneLoaderSystem a Handle<GltfSceneAsset> from this format can be attached to an entity in ECS, and the system will then load the full scene using the given entity as the root node of the scene hierarchy.

See GltfSceneOptions for more information about the load options.